inky-3/inky-dashboard/requirements.txt
spotipy
flask
requests
google-auth
google-auth-oauthlib
google-api-python-client
Pillow
evdev
gpiod
gpiodevice
qrcode[pil]

# pip install -r requirements.txt
